Refactor folder setup for ease of use
Previously the frontend and backend were separate repositories with
their own docker-compose.yml files. This made it awkward to run,
having to run two commands instead of one to get it working. This
change also makes it easier to deploy the project to ECS, as all
the files are in one place.